+++ src/main/java/callprotector/spring/domain/callchat/controller/CallChatLogController.java
@@ -2,8 +2,8 @@ package callprotector.spring.domain.callchat.controller;
 
 import callprotector.spring.domain.callchat.dto.response.CallChatLogResponseDTO;
 import callprotector.spring.domain.callchat.entity.CallChatSession;
-import callprotector.spring.domain.chat.service.CallChatLogService;
-import callprotector.spring.domain.chat.service.CallChatSessionService;
+import callprotector.spring.domain.callchat.service.CallChatLogService;
+import callprotector.spring.domain.callchat.service.CallChatSessionService;
 import callprotector.spring.global.annotation.UserId;
 import callprotector.spring.global.apiPayload.ApiResponse;
 import lombok.RequiredArgsConstructor;
