+++ src/main/java/callprotector/spring/domain/callchat/controller/CallChatStreamController.java
@@ -3,8 +3,8 @@ package callprotector.spring.domain.callchat.controller;
 import callprotector.spring.domain.callsttlog.entity.CallSttLog;
 import callprotector.spring.domain.callsttlog.service.CallSttLogService;
 import callprotector.spring.domain.callchat.entity.CallChatSession;
-import callprotector.spring.domain.chat.service.CallChatLogService;
-import callprotector.spring.domain.chat.service.CallChatSessionService;
+import callprotector.spring.domain.callchat.service.CallChatLogService;
+import callprotector.spring.domain.callchat.service.CallChatSessionService;
 import callprotector.spring.global.security.TokenProvider;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
